+++ artemis-server/src/main/java/org/apache/activemq/artemis/core/replication/ReplicationManager.java
@@ -288,19 +288,19 @@ public final class ReplicationManager implements ActiveMQComponent {
     * @param storeName
     * @param pageNumber
     */
-   public void pageClosed(final SimpleString storeName, final int pageNumber) {
+   public void pageClosed(final SimpleString storeName, final long pageNumber) {
       if (enabled) {
          sendReplicatePacket(new ReplicationPageEventMessage(storeName, pageNumber, false));
       }
    }
 
-   public void pageDeleted(final SimpleString storeName, final int pageNumber) {
+   public void pageDeleted(final SimpleString storeName, final long pageNumber) {
       if (enabled) {
          sendReplicatePacket(new ReplicationPageEventMessage(storeName, pageNumber, true));
       }
    }
 
-   public void pageWrite(final PagedMessage message, final int pageNumber) {
+   public void pageWrite(final PagedMessage message, final long pageNumber) {
       if (enabled) {
          sendReplicatePacket(new ReplicationPageWriteMessage(message, pageNumber));
       }
